backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2
    # via starlette
click==8.1.7
    # via uvicorn
fastapi==0.115.5
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.10
    # via anyio
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic==2.10.1
    # via fastapi
pydantic-core==2.27.1
    # via pydantic
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.32.1
    # via -r requirements.in

# next is custom packages
# pretty printing objects
devtools

# jwt and passwords
python-jose
bcrypt
passlib[bcrypt]

# for async db interaction
aiopg

# for requests to dumper
requests
